refactor: 주요 컴포넌트 파일을 lib 폴더 하위로 이동
