--- src/Container/style.css
@@ -1,4 +1,0 @@
-.container {
-    display: flex;
-    justify-content: center;
-}
