--- src/Flex/style.css
@@ -1,3 +1,0 @@
-.flex {
-  display: flex;
-}
