--- src/Grid/style.css
@@ -1,3 +1,0 @@
-.grid {
-  display: grid;
-}
