+++ src/index.tsx
@@ -1,5 +1,5 @@
-import Container from "./Container";
-import Flex from "./Flex";
-import Grid from "./Grid";
+import Container from "./lib/Container";
+import Flex from "./lib/Flex";
+import Grid from "./lib/Grid";
 
 export { Container, Flex, Grid };
