+++ src/lib/Grid/index.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import "./style.css";
 
 interface GridProps extends React.PropsWithChildren {
   rows: number;
@@ -9,14 +8,11 @@ interface GridProps extends React.PropsWithChildren {
 
 export default function Grid({ rows, columns, gap, children }: GridProps) {
   const style = {
+    display: "grid",
     gridTemplateColumns: `repeat(${columns}, 1fr)`,
     gridTemplateRows: `repeat(${rows}, 1fr)`,
     gap: `${gap}px`,
   };
 
-  return (
-    <div className="grid" style={style}>
-      {children}
-    </div>
-  );
+  return <div style={style}>{children}</div>;
 }
